buidl: refactor, main function

+++ scripts/buidl.sh
@@ -1,38 +1,37 @@
 #!/usr/bin/env bash
 # Script for building a host with injected json data
+# Dependencies: Nix package manager, nix-command 
 
 set -o pipefail
 trap cleanup EXIT
 trap 'cleanup && exit 0' SIGINT
 
-# Default values
+# Default argument values
 verbose=false
 output_path="./result"
 
 # Do not change, this path is also hard-coded in flake.nix
-data_nix="/tmp/data.nix" 
+data_nix="/tmp/data.nix"
 
 # Cleanup which will be executed at exit
 cleanup() {
-  if [ -f "$data_nix" ]; then
-    rm "$data_nix"
+  [[ -f "$data_nix" ]] && rm "$data_nix"
+}
+
+check_dependecies() {
+  if ! command -v nix-instantiate >/dev/null 2>&1; then
+    echo "error: 'nix-instantiate' command not found. Please install the Nix package manager."
+    echo "for installation instructions, visit: https://nixos.org/download.html"
+    exit 1
+  else
+    if ! command -v nix >/dev/null 2>&1; then
+      echo "error: 'nix' command not found. Please install the nix-command."
+      echo "you can install the nix-command by running 'nix-env -iA nixpkgs.nix'."
+      exit 1
+    fi
   fi
 }
 
-# Check dependencies 
-if ! command -v nix-instantiate >/dev/null 2>&1; then
-  echo "error: 'nix-instantiate' command not found. Please install the Nix package manager."
-  echo "for installation instructions, visit: https://nixos.org/download.html"
-  exit 1
-else
-  if ! command -v nix >/dev/null 2>&1; then
-    echo "error: 'nix' command not found. Please install the nix-command."
-    echo "you can install the nix-command by running 'nix-env -iA nixpkgs.nix'."
-    exit 1
-  fi
-fi
-
-# Function to display help message
 display_usage() {
   cat <<USAGE
 Usage: $0 [options] [json_data]
@@ -64,63 +63,45 @@ Examples:
 USAGE
 }
 
-# Parse command line arguments
-while [[ "$#" -gt 0 ]]; do
-  case $1 in
-    -b|--base)
-      hostname="$2"
-      shift 2 ;;
-    -o|--output)
-      output_path="$2"
-      shift 2 ;;
-    -v|--verbose)
-      verbose=true
-      shift ;;
-    -h|--help)
-      display_usage
-      exit 0 ;;
-    *)
-      # Check if argument is JSON data
-      if [[ "$1" =~ ^\{.*\}$ ]]; then
-        json_data="$1"
-      else
-        echo "error: unknown option -- '$1'"
-        echo "try '--help' for more information."
-        exit 1
-      fi
-      shift ;;
-  esac
-done
+parse_arguments() {
+  while [[ "$#" -gt 0 ]]; do
+    case $1 in
+      -b|--base)
+        hostname="$2"
+        shift 2 ;;
+      -o|--output)
+        output_path="$2"
+        shift 2 ;;
+      -v|--verbose)
+        verbose=true
+        shift ;;
+      -h|--help)
+        display_usage
+        exit 0 ;;
+      *)
+        # Check if argument is JSON data
+        if [[ "$1" =~ ^\{.*\}$ ]]; then
+          json_data="$1"
+        else
+          echo "error: unknown option -- '$1'"
+          echo "try '--help' for more information."
+          exit 1
+        fi
+        shift ;;
+    esac
+  done
+  # Check that base configuration has been set
+  if [[ -z $hostname ]]; then
+    echo "error: base configuration is required."
+    echo "try '--help' for more information."
+    exit 1
+  fi
+}
 
-# Default flags for 'nix build' command
-nix_flags=(
-  --accept-flake-config
-  --extra-experimental-features 'nix-command flakes'
-  --impure
-  --no-warn-dirty
-  --out-link "$output_path"
-)
+create_data_nix() {
+  local json_data="$1"
+  local data_nix="$2"
 
-# Append '--show-trace' and '--debug' if verbose flag is true
-if [ "$verbose" = true ]; then
-  nix_flags+=("--show-trace" "--debug" )
-fi
-
-# Read JSON data from stdin if it's not provided as an argument
-if [ -z "$json_data" ] && ! tty -s && [ -p /dev/stdin ]; then
-  # Read JSON data from stdin
-  json_data=$(</dev/stdin)
-fi
-
-# Check that base configuration has been set
-if [[ -z $hostname ]]; then
-  echo "error: base configuration is required."
-  echo "try '--help' for more information."
-  exit 1
-fi
-
-# Create data.nix from JSON data
-if [[ -n $json_data ]]; then
   # Escape double quotes
   esc_json_data="${json_data//\"/\\\"}"
 
@@ -134,30 +115,78 @@ if [[ -n $json_data ]]; then
   $hostname = $nix_expr;
 }
 EOF
-else
-  # Precaution
-  cleanup
-fi
+}
 
-# Run nix build command
-output=$(nix build .#"$hostname" "${nix_flags[@]}")
-if [[ $output == *"error: could not find a flake.nix file"* ]]; then
-  if [ "$verbose" = true ]; then echo "fetching from github:ponkila/nixobolus"; fi
-  nix build github:ponkila/nixobolus#"$hostname" "${nix_flags[@]}" || exit 1
-fi
+run_nix_build() {
+  local hostname="$1"
+  local output_path="$2"
+  local verbose="$3"
 
-# Display injected data if verbose is true
-if [ "$verbose" = true ] && [ -f "$data_nix" ]; then
-  # Replace newlines with spaces, removes consecutive spaces and trailing space
-  echo "injected data: '$(cat $data_nix | tr '\n' ' ' | tr -s ' ' | sed 's/ $//')'"
-fi
+  # Default flags for the 'nix build' command
+  declare -a nix_flags=(
+    --accept-flake-config
+    --extra-experimental-features 'nix-command flakes'
+    --impure
+    --no-warn-dirty
+    --out-link "$output_path"
+  )
 
-# Print the real paths of the symlinks
-for symlink in "$output_path"/*; do
-  real_path=$(readlink -f "$symlink")
-  if [ "$verbose" = true ]; then 
-    echo created symlink: \'"$symlink > $real_path"\'
-  else
-    echo "$real_path"
+  # Append '--show-trace' and '--debug' if verbose flag is true
+  [[ "$verbose" = true ]] && nix_flags+=("--show-trace" "--debug")
+  
+  # Execute the 'nix build' command
+  output=$(nix build .#"$hostname" "${nix_flags[@]}")
+
+  # Fallback to fetching from GitHub if the flake.nix file is not found
+  if [[ $output == *"error: could not find a flake.nix file"* ]]; then
+    if [ "$verbose" = true ]; then echo "fetching from github:ponkila/nixobolus"; fi
+    nix build github:ponkila/nixobolus#"$hostname" "${nix_flags[@]}" || exit 1
   fi
-done
+}
+
+print_output() {
+  local output_path="$1"
+  local data_nix="$2"
+  local verbose="$3"
+
+  # Display injected data if verbose is true
+  if [ "$verbose" = true ] && [ -f "$data_nix" ]; then
+    # Replaces newlines with spaces, removes consecutive spaces and trailing space
+  echo "injected data: '$(< "$data_nix" tr '\n' ' ' | tr -s ' ' | sed 's/ $//')'"
+  fi
+
+  # Print the real paths of the symlinks
+  for symlink in "$output_path"/*; do
+    real_path=$(readlink -f "$symlink")
+    if [ "$verbose" = true ]; then 
+      echo created symlink: \'"$symlink > $real_path"\'
+    else
+      echo "$real_path"
+    fi
+  done
+}
+
+main() {
+  # Check necessary dependencies to run the script
+  check_dependecies
+
+  # Parse and validate command line arguments
+  parse_arguments "$@"
+
+  # Read JSON data from stdin if it exists and is not provided as an argument
+  if [ -z "$json_data" ] && ! tty -s && [ -p /dev/stdin ]; then
+    json_data=$(</dev/stdin)
+  fi
+
+  # If JSON data is provided, create 'data.nix' from it
+  # This file will be automatically imported by the flake if it exists
+  [[ -n $json_data ]] && create_data_nix "$json_data" "$data_nix"
+
+  # Run the 'nix build' command and fallback to fetching from GitHub if the flake is not found
+  run_nix_build "$hostname" "$output_path" $verbose
+
+  # Display additional output, including injected data and created symlinks
+  print_output "$output_path" "$data_nix" $verbose
+}
+
+main "$@"
